quickstart: Update the required vCPU numbers to 4
While preparing the `v0.3.0` release, we've noticed that using a VM with
2 vCPUs would lead to:
```
Name:           cc-operator-controller-manager-79797456f6-spmss
Namespace:      confidential-containers-system
...
Events:
  Type     Reason            Age    From               Message
  ----     ------            ----   ----               -------
  Warning  FailedScheduling  4m12s  default-scheduler  0/1 nodes are available: 1 Insufficient cpu. preemption: 0/1 nodes are available: 1 No preemption victims found for incoming pod.
```

And this is *NOT* something introduced between `v0.2.0` and `v0.3.0`, as
it also happen with the previous release.

For now, let's update the documentation accordingly and revisit this
after the release in case we need to really rely on deploying in nodes
with 2 vCPUs.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>